refactored logic into main.service and fixed on_post
